refactor About
