+++ client/src/components/MainContent/MainContent.tsx
@@ -1,6 +1,6 @@
 import styles from "./MainContent.module.css"
 import ContentCard from "@/components/MainContent/ContentCard/ContentCard";
-import About from "@/components/MainContent/About";
+import About from "@/components/About/About";
 import Skills from "@/components/MainContent/Skills";
 import Education from "@/components/MainContent/Education/Education";
 import Experience from "@/components/MainContent/Experience/Experience";
